+++ src/PhpSpreadsheet/Comment.php
@@ -96,7 +96,7 @@ class Comment implements IComparable
      *
      * @param string $author
      *
-     * @return Comment
+     * @return $this
      */
     public function setAuthor($author)
     {
@@ -120,7 +120,7 @@ class Comment implements IComparable
      *
      * @param RichText $pValue
      *
-     * @return Comment
+     * @return $this
      */
     public function setText(RichText $pValue)
     {
@@ -144,7 +144,7 @@ class Comment implements IComparable
      *
      * @param string $width
      *
-     * @return Comment
+     * @return $this
      */
     public function setWidth($width)
     {
@@ -168,7 +168,7 @@ class Comment implements IComparable
      *
      * @param string $value
      *
-     * @return Comment
+     * @return $this
      */
     public function setHeight($value)
     {
@@ -192,7 +192,7 @@ class Comment implements IComparable
      *
      * @param string $value
      *
-     * @return Comment
+     * @return $this
      */
     public function setMarginLeft($value)
     {
@@ -216,7 +216,7 @@ class Comment implements IComparable
      *
      * @param string $value
      *
-     * @return Comment
+     * @return $this
      */
     public function setMarginTop($value)
     {
@@ -240,7 +240,7 @@ class Comment implements IComparable
      *
      * @param bool $value
      *
-     * @return Comment
+     * @return $this
      */
     public function setVisible($value)
     {
@@ -264,7 +264,7 @@ class Comment implements IComparable
      *
      * @param string $alignment see Style\Alignment::HORIZONTAL_*
      *
-     * @return Comment
+     * @return $this
      */
     public function setAlignment($alignment)
     {