+++ src/PhpSpreadsheet/Style/Conditional.php
@@ -12,6 +12,7 @@ class Conditional implements IComparable
     const CONDITION_CONTAINSTEXT = 'containsText';
     const CONDITION_EXPRESSION = 'expression';
     const CONDITION_CONTAINSBLANKS = 'containsBlanks';
+    const CONDITION_NOTCONTAINSBLANKS = 'notContainsBlanks';
 
     // Operator types
     const OPERATOR_NONE = '';
@@ -93,7 +94,7 @@ class Conditional implements IComparable
      *
      * @param string $pValue Condition type, see self::CONDITION_*
      *
-     * @return Conditional
+     * @return $this
      */
     public function setConditionType($pValue)
     {
@@ -117,7 +118,7 @@ class Conditional implements IComparable
      *
      * @param string $pValue Conditional operator type, see self::OPERATOR_*
      *
-     * @return Conditional
+     * @return $this
      */
     public function setOperatorType($pValue)
     {
@@ -141,7 +142,7 @@ class Conditional implements IComparable
      *
      * @param string $value
      *
-     * @return Conditional
+     * @return $this
      */
     public function setText($value)
     {
@@ -165,7 +166,7 @@ class Conditional implements IComparable
      *
      * @param bool $value
      *
-     * @return Conditional
+     * @return $this
      */
     public function setStopIfTrue($value)
     {
@@ -189,7 +190,7 @@ class Conditional implements IComparable
      *
      * @param string[] $pValue Condition
      *
-     * @return Conditional
+     * @return $this
      */
     public function setConditions($pValue)
     {
@@ -206,7 +207,7 @@ class Conditional implements IComparable
      *
      * @param string $pValue Condition
      *
-     * @return Conditional
+     * @return $this
      */
     public function addCondition($pValue)
     {
@@ -230,7 +231,7 @@ class Conditional implements IComparable
      *
      * @param Style $pValue
      *
-     * @return Conditional
+     * @return $this
      */
     public function setStyle(Style $pValue = null)
     {