+++ src/PhpSpreadsheet/Style/Protection.php
@@ -86,7 +86,7 @@ class Protection extends Supervisor
      *
      * @throws PhpSpreadsheetException
      *
-     * @return Protection
+     * @return $this
      */
     public function applyFromArray(array $pStyles)
     {
@@ -123,7 +123,7 @@ class Protection extends Supervisor
      *
      * @param string $pValue see self::PROTECTION_*
      *
-     * @return Protection
+     * @return $this
      */
     public function setLocked($pValue)
     {
@@ -156,7 +156,7 @@ class Protection extends Supervisor
      *
      * @param string $pValue see self::PROTECTION_*
      *
-     * @return Protection
+     * @return $this
      */
     public function setHidden($pValue)
     {