+++ tests/data/Cell/SetValueExplicitException.php
@@ -1,0 +1,10 @@
+<?php
+
+use PhpOffice\PhpSpreadsheet\Cell\DataType;
+
+return [
+    [
+        'XYZ',
+        DataType::TYPE_NUMERIC,
+    ],
+];